[Feature] adicionar estrutura inicial do aplicativo de esportes com rotas, views e templates
