+++ apps/esportemuz/templates/esportemuz/pages/campeonato.html
@@ -1,0 +1,5 @@
+{% extends "esportemuz/base.html" %}
+
+{% block content %}
+<h1>Campeonato</h1>
+{% endblock content %}
