+++ apps/esportemuz/templates/esportemuz/pages/equipe.html
@@ -1,0 +1,5 @@
+{% extends "esportemuz/base.html" %}
+
+{% block content %}
+<h1>Equipe</h1>
+{% endblock content %}
